Remove code doing stack optimization with PHI nodes.

+++ evmjit/libevmjit/RuntimeManager.h
@@ -50,7 +50,7 @@ public:
 	static llvm::StructType* getRuntimeType();
 	static llvm::StructType* getRuntimeDataType();
 
-	void checkStackLimit(size_t _max, int _diff);
+	void checkStackLimit(size_t _max, ssize_t _diff);
 
 private:
 	llvm::Value* getPtr(RuntimeData::Index _index);
